Add config with HSTS for production (#102)
Currently, the UI is served via both http and both https. In the former case, a warning will appear and all data will be sent unencrypted over the network. Also, LS Login will not work because it requires the https based client URL.

This can be fixed by configuring the API gateway to send a redirect for http requests and/or by sending an HSTS (HTTP Strict Transport Security) header with the application. This PR implements the latter. It adds a serve.json configuration file that adds a HSTS header, but can also be used for [other purposes](https://github.com/vercel/serve-handler#options).