Some comments added to the code. (#115)
I added detailed comments in TSDoc format to the functions in the TS files. I wrote only one sentence definitions for TSX components. Since it will be a long-term job to cover all the code line by line, it will be more efficient to update the code as I work on it in the future.