Fetch user data from userinfo instead of token (#121)
Until recently, LS AAI provided the user data we are interested in (email and name) as part of its OIDC access token. This was very convenient for us since we only needed to deserialize the token to get the info. However, recently this was changed and the token does not contain this information any more. Instead, it must be fetched from the userinfo endpoint.

This PR adapts the user portal accordingly. Note that the auth adapter used by the API gateway also needs to be adapted.